Moved async wait directory to own file
